requirements.txt
asgiref==3.9.1
Django==5.2.6
django-debug-toolbar==6.0.0
djangorestframework==3.16.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
